+++ config/design-laravel-kit.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'views_namespace' => 'design-laravel-kit',
+    'component_prefix' => 'italia',
+    'id_prefix' => 'dlk',
+    'assets_path' => 'vendor/design-laravel-kit',
+];